+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 --------------------------------
 - ADDED: Ability to change duration, interpolation and start delay of `AnimatorViewChangeHandler`.
 
+- ADDED: `jumpToRoot(direction)`.
+
 -Simple Stack 1.9.2 (2018-06-01)
 --------------------------------
 - MINOR CHANGE: `DefaultStateChanger` uses `FadeViewChanger` for `StateChange.REPLACE` by default, instead of `NoOpViewChangeHandler`.